Disable keep-alives in shutdown hander
If a persistent connection is before shutdown, the client will
continue to be able to make requests to the server and block
the shutdown because the `ConnState` never changes to `http.StateClosed`.
This disables keep-alives on the server during the shutdown routine,
which is recommended in the net/http docs for servers in the process
of shutting down.